+++ Source/Libs/ManagedIrbis5/Source/Connectivity/SyncConnection.cs
@@ -225,7 +225,7 @@ namespace ManagedIrbis
                     Debug.WriteLine (exception.Message);
                     if (perfRecord is not null)
                     {
-                        perfRecord.ElapsedTime = stopwatch!.ElapsedMilliseconds;
+                        perfRecord.ElapsedTime = (int) stopwatch!.ElapsedMilliseconds;
                         perfRecord.ErrorMessage = exception.Message;
                         _performanceCollector!.Collect (perfRecord);
                     }
@@ -237,7 +237,7 @@ namespace ManagedIrbis
                 {
                     if (perfRecord is not null)
                     {
-                        perfRecord.ElapsedTime = stopwatch!.ElapsedMilliseconds;
+                        perfRecord.ElapsedTime = (int) stopwatch!.ElapsedMilliseconds;
                         perfRecord.ErrorMessage = "No response";
                         _performanceCollector!.Collect (perfRecord);
                     }
@@ -255,7 +255,7 @@ namespace ManagedIrbis
                 result.Parse();
                 if (perfRecord is not null)
                 {
-                    perfRecord.ElapsedTime = stopwatch!.ElapsedMilliseconds;
+                    perfRecord.ElapsedTime = (int) stopwatch!.ElapsedMilliseconds;
                     perfRecord.IncomingSize = result.AnswerSize;
                     _performanceCollector!.Collect (perfRecord);
                 }
